requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
google-cloud-bigquery>=3.13.0
db-dtypes>=1.1.1
pyarrow>=14.0.0